refactor(*): remove redux stuff
