+++ extension-mobx/src/components/Stream/spec.js
@@ -1,4 +1,4 @@
-import Stream from './presenter';
+import Stream from '../Stream';
 import { shallow } from 'enzyme';
 
 describe('Stream', () => {
