--- extension-mobx/src/constants/actionTypes.js
@@ -1,3 +1,0 @@
-export const ME_SET = 'ME_SET';
-export const TRACKS_SET = 'TRACKS_SET';
-export const TRACK_PLAY = 'TRACK_PLAY';
