--- extension-mobx/src/stores/configureStore.js
@@ -1,15 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import createLogger from 'redux-logger';
-import thunk from 'redux-thunk';
-import { browserHistory } from 'react-router';
-import { routerMiddleware } from 'react-router-redux';
-import rootReducer from '../reducers/index';
-
-const logger = createLogger();
-const router = routerMiddleware(browserHistory);
-
-const createStoreWithMiddleware = applyMiddleware(thunk, router)(createStore);
-
-export default function configureStore(initialState) {
-  return createStoreWithMiddleware(rootReducer, initialState);
-}
